refactor: change folder names and move exported code outside internal scope
