+++ src/lib/components/ui/form-text-field/form-text-field.svelte
@@ -1,0 +1,25 @@
+<script
+	lang="ts"
+	module
+></script>
+
+<!-- <script lang="ts">
+	import { TextField } from '$components/ui/text-field';
+	import { TextFieldLabel } from '../text-field-label';
+</script>
+
+<div class="relative">
+	<TextFieldLabel for="test">hello</TextFieldLabel>
+	<TextField
+		variant="filled"
+		bind:value
+		size="lg"
+		color="secondary"
+		id="test"
+		aria-describedby="test-desc"
+		aria-invalid={isInvalid}
+		{required}
+		{disabled}
+		{placeholder}
+	/>
+</div> -->
