+++ src/lib/components/ui/form-text-field/index.ts
@@ -1,0 +1,1 @@
+export { default as FormTextField } from '$components/ui/form-text-field/form-text-field.svelte';
