--- src/lib/components/ui/forms/button/index.ts
@@ -1,4 +1,0 @@
-export {
-	default as Button,
-	type ButtonProps,
-} from '$components/ui/forms/button/button.svelte';
