--- src/lib/components/ui/forms/descriptions/label/index.ts
@@ -1,4 +1,0 @@
-export {
-	default as Label,
-	type LabelProps,
-} from '$components/ui/forms/descriptions/label/label.svelte';
