--- src/lib/components/ui/forms/fields/text-field/text-field.svelte
@@ -1,10 +1,0 @@
-<script
-	lang="ts"
-	module
-></script>
-
-<script lang="ts">
-	import { TextFieldBase } from '$components/ui/forms/fields/text-field/base';
-</script>
-
-<TextFieldBase></TextFieldBase>
