+++ src/routes/+page.svelte
@@ -1,33 +1,61 @@
 <script lang="ts">
-	import { Label } from '../lib/components/ui/forms/descriptions/label';
-	import TextFieldBase from '../lib/components/ui/forms/fields/text-field/base/text-field-base.svelte';
+	import { TextField } from '../lib/components/ui/text-field';
+	import { TextFieldLabel } from '../lib/components/ui/text-field-label';
 
-	let placeholder = '';
-
+	let placeholder = undefined;
+	let disabled = false;
+	let required = true;
 	let value = $state<string | undefined>();
-
-	const isPlaceholderShown = $derived(
-		(value === undefined || value === '') && placeholder !== undefined,
-	);
+	let isInvalid = $state(false);
 </script>
 
 <div class="mb-20 w-full">d</div>
 <div class="mx-12 flex items-center gap-2">
+	<!-- to the form we will add novalidate to prevent browser defaults -->
+	<!--  -->
+	<!-- FormTextField -->
+	<!--  -->
+	<!-- FormControl -->
 	<div
-		class="group relative size-fit space-y-2"
-		data-placeholder-shown={isPlaceholderShown}
+		class="group my-3 h-fit w-full space-y-1"
+		data-input-placeholder-shown={((value === undefined || value === '') &&
+			placeholder !== undefined) ||
+			undefined}
+		data-input-has-value={(value !== undefined && value !== '') || undefined}
+		data-input-disabled={disabled || undefined}
+		data-input-invalid={isInvalid || undefined}
+		data-input-required={required || undefined}
 	>
-		<Label
-			for="test"
-			class="-placeholder-shown:text-red-500 absolute top-0 left-0 mt-3 min-h-11 pt-3.5 pl-5 transition-all duration-200 group-data-[placeholder-shown=false]:mt-0 group-data-[placeholder-shown=false]:pt-1.5 peer-focus-visible:mt-0 peer-focus-visible:pt-1.5"
-			>hello</Label
+		<!-- FormTextField -->
+		<!-- FormField -->
+		<div class="relative">
+			<TextFieldLabel
+				for="test"
+				variant="filled"
+				color="secondary"
+				size="lg">hello</TextFieldLabel
+			>
+			<TextField
+				variant="filled"
+				bind:value
+				size="lg"
+				color="secondary"
+				id="test"
+				aria-describedby="test-desc"
+				aria-invalid={isInvalid}
+				{required}
+				{disabled}
+				{placeholder}
+			/>
+		</div>
+		<!--  -->
+		<!-- Form Description can change to be a description or an error message -->
+		<p
+			id="test-desc"
+			class="group-data-[input-invalid]:text-destructive pl-5"
 		>
-		<TextFieldBase
-			variant="filled"
-			bind:value
-			size="lg"
-			id="test"
-			{placeholder}
-		/>
+			desc
+		</p>
 	</div>
+	<!--  -->
 </div>
